+++ src/agentic_news/agents/fact_checking_agent.py
@@ -1,5 +1,19 @@
-
-
+_fact_checking_prompt: str = "You have been provided the results of some research. Compare and contrast the results of the research for fact-checking purposes and provide the different ideas and cited sources for each that support and go against the results"
 
 def fact_checking_agent(research_results: str) -> str:
-    return ""
+    """Takes in the results of the research as a string and then returns a string containing supporting and refuting evidences of the ideas presented in the research.
+    Args:
+        research_results <str>: input research results
+    Returns:
+        <str>: A summary of the supporting and refuting ideas within the research and the citations
+    """
+    client = genai.Client()
+    # create the research agent
+    interaction = client.interactions.create(
+        model="gemini-3.1-flash",
+        config={
+            "system_instruction": _fact_checking_prompt
+        },
+        contents=research_results
+    )
+    return interaction.output_text